+++ crankshaft-docker/src/container.rs
@@ -16,6 +16,7 @@ use bollard::container::RemoveContainerOptions;
 use bollard::container::StartContainerOptions;
 use bollard::container::UploadToContainerOptions;
 use bollard::container::WaitContainerOptions;
+use bollard::secret::ContainerWaitResponse;
 use futures::TryStreamExt as _;
 use tokio_stream::StreamExt as _;
 use tracing::debug;
@@ -112,6 +113,8 @@ impl Container {
             .map_err(Error::Docker)?
             .output;
 
+        debug!("starting container `{name}`", name = self.name);
+
         // Start the container.
         self.client
             .start_container(&self.name, None::<StartContainerOptions<String>>)
@@ -148,23 +151,24 @@ impl Container {
             .map_err(Error::Docker)?;
 
         // Wait for the container to be completed.
+
+        debug!("waiting for container `{name}` to exit", name = self.name);
         let mut wait_stream = self
             .client
             .wait_container(&self.name, None::<WaitContainerOptions<String>>);
 
         let mut exit_code = None;
         if let Some(result) = wait_stream.next().await {
-            let response = result.map_err(Error::Docker)?;
-            if let Some(error) = response.error {
-                return Err(Error::Message(format!(
-                    "failed to wait for Docker task to complete: {error}",
-                    error = error
-                        .message
-                        .expect("Docker reported an error without a message")
-                )));
+            match result {
+                // Bollard turns non-zero exit codes into wait errors, so check for both
+                Ok(ContainerWaitResponse {
+                    status_code: code, ..
+                })
+                | Err(bollard::errors::Error::DockerContainerWaitError { code, .. }) => {
+                    exit_code = Some(code);
+                }
+                Err(e) => return Err(e.into()),
             }
-
-            exit_code = Some(response.status_code);
         }
 
         if exit_code.is_none() {
@@ -186,7 +190,8 @@ impl Container {
 
         #[cfg(unix)]
         let output = Output {
-            status: ExitStatus::from_raw(exit_code.unwrap() as i32),
+            // See WEXITSTATUS from wait(2) to explain the shift
+            status: ExitStatus::from_raw((exit_code.unwrap() as i32) << 8),
             stdout,
             stderr,
         };