+++ crankshaft-engine/src/service/runner/backend.rs
@@ -2,12 +2,12 @@
 
 use std::fmt::Debug;
 use std::process::Output;
-use std::sync::Arc;
 
 use async_trait::async_trait;
 use eyre::Result;
 use futures::future::BoxFuture;
 use nonempty::NonEmpty;
+use tokio::sync::oneshot;
 use tokio_util::sync::CancellationToken;
 
 use crate::Task;
@@ -24,16 +24,15 @@ pub trait Backend: Debug + Send + Sync + 'static {
 
     /// Runs a task in a backend.
     ///
-    /// The `started` callback is called when an execution of the task has
-    /// started; the parameter is the index of the execution into the task's
-    /// executions collection.
+    /// The optional `started` channel is notified when the first execution of
+    /// the task has started.
     // TODO: use a representation of task output that isn't based on
     // `std::process::Output` that would allow us to write stdout/stderror to a file
     // instead of buffering it all in memory
     fn run(
         &self,
         task: Task,
-        started: Arc<dyn Fn(usize) + Send + Sync + 'static>,
+        started: Option<oneshot::Sender<()>>,
         token: CancellationToken,
     ) -> Result<BoxFuture<'static, Result<NonEmpty<Output>>>>;
 }